Add files via upload
This script estimate change points across a time series using SEGMENTED model.

npsi allows to set up the number og change points to be detected
The script controls if the model is struggling to estimate breakpoints properly, likely due to overfitting or excessive breakpoint constraints based on the nspi value.
The mode can modify the number of iterations (in this case = 50)
The script identifies the most significant change point from all change points detected based on the magnitude of NDVI shifts before and after each breakpoint. This ensures that we are extracting the most impactful structural shift in each time series.